+++ Tortuga.Chain/Shared/Tests/Aggregation/ComplexAggregateTests.cs
@@ -65,4 +65,78 @@ public class ComplexAggregateTests : TestBase
 			Release(dataSource);
 		}
 	}
+
+	[DataTestMethod, TablesAndViewData(DataSourceGroup.All)]
+	public void AggregateObject(string dataSourceName, DataSourceType mode, string tableName)
+	{
+		var dataSource = DataSource(dataSourceName, mode);
+		try
+		{
+			//PostgreSQL is case sensitive, so we need to ensure we're using the correct name.
+			var table = dataSource.DatabaseMetadata.GetTableOrViewFromClass<Employee>();
+			var ekColumnName = table.Columns["EmployeeKey"].SqlName;
+			var gColumnName = table.Columns["Gender"].SqlName;
+
+			var result = dataSource.From<Employee>(Filter).AsAggregate<EmployeeReport>().ToObject().Execute();
+		}
+		finally
+		{
+			Release(dataSource);
+		}
+	}
+
+	[DataTestMethod, TablesAndViewData(DataSourceGroup.All)]
+	public void AggregateObject_WithGroup(string dataSourceName, DataSourceType mode, string tableName)
+	{
+		var dataSource = DataSource(dataSourceName, mode);
+		try
+		{
+			//PostgreSQL is case sensitive, so we need to ensure we're using the correct name.
+			var table = dataSource.DatabaseMetadata.GetTableOrViewFromClass<Employee>();
+			var ekColumnName = table.Columns["EmployeeKey"].SqlName;
+			var gColumnName = table.Columns["Gender"].SqlName;
+
+			var result = dataSource.From<Employee>(Filter).AsAggregate<GroupedEmployeeReport>().ToCollection().Execute();
+		}
+		finally
+		{
+			Release(dataSource);
+		}
+	}
+
+	public class GroupedEmployeeReport
+	{
+#if POSTGRESQL
+		const string ekColumnName = "employeekey";
+#else
+		const string ekColumnName = "EmployeeKey";
+#endif
+
+		[AggregateColumn(AggregateType.Min, "EmployeeKey")]
+		public int MinEmployeeKey { get; set; }
+
+		[AggregateColumn(AggregateType.Max, "EmployeeKey")]
+		public int MaxEmployeeKey { get; set; }
+
+		[AggregateColumn(AggregateType.Count, "EmployeeKey")]
+		public int CountEmployeeKey { get; set; }
+
+		[GroupByColumn]
+		public string Gender { get; set; }
+
+		[CustomAggregateColumn($"Max({ekColumnName}) - Min({ekColumnName})")]
+		public int Range { get; set; }
+	}
+
+	public class EmployeeReport
+	{
+		[AggregateColumn(AggregateType.Min, "EmployeeKey")]
+		public int MinEmployeeKey { get; set; }
+
+		[AggregateColumn(AggregateType.Max, "EmployeeKey")]
+		public int MaxEmployeeKey { get; set; }
+
+		[AggregateColumn(AggregateType.Count, "EmployeeKey")]
+		public int CountEmployeeKey { get; set; }
+	}
 }