+++ Tortuga.Chain/Tortuga.Chain.Core/Aggregates/GroupByColumnAttribute.cs
@@ -6,25 +6,38 @@
 /// </summary>
 /// <seealso cref="Attribute" />
 [AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
-public class GroupByColumnAttribute : Attribute
+public class GroupByColumnAttribute : BaseAggregateAttribute
 {
 	/// <summary>
 	/// Initializes a new instance of the <see cref="GroupByColumnAttribute"/> class.
 	/// </summary>
 	/// <param name="sourceColumnName">Name of the source column. If null, uses the property name.</param>
-	/// <param name="groupingOrder">The order to apply the group by columns. If null, uses the order the properties appear in reflection.</param>
-	public GroupByColumnAttribute(string? sourceColumnName = null, int? groupingOrder = null)
+	public GroupByColumnAttribute(string? sourceColumnName = null)
 	{
 		SourceColumnName = sourceColumnName;
-		GroupingOrder = groupingOrder ?? int.MaxValue;
 	}
 
 	/// <summary>
-	/// The order to apply the group by columns. If tied, uses the order the properties appear in reflection.
+	/// Initializes a new instance of the <see cref="GroupByColumnAttribute"/> class.
 	/// </summary>
-	/// <value>The grouping order.</value>
-	/// <remarks>Defaults to Int32.MaxValue</remarks>
-	public int GroupingOrder { get; }
+	/// <param name="sourceColumnName">Name of the source column. If null, uses the property name.</param>
+	/// <param name="order">The order to apply the group by columns. If null, uses the order the properties appear in reflection.</param>
+	public GroupByColumnAttribute(string? sourceColumnName, int order)
+	{
+		SourceColumnName = sourceColumnName;
+		Order = order;
+		GroupBy = true;
+	}
+
+	/// <summary>
+	/// Initializes a new instance of the <see cref="GroupByColumnAttribute"/> class.
+	/// </summary>
+	/// <param name="order">The order to apply the group by columns. If null, uses the order the properties appear in reflection.</param>
+	public GroupByColumnAttribute(int order)
+	{
+		Order = order;
+		GroupBy = true;
+	}
 
 	/// <summary>
 	/// Gets the name of the source column to be provided to the aggregate function.